services/atlablog/requirements.txt
sanic==0.1.7
aiohttp==1.0.5
peewee-async==0.5.5
aiopg==0.12.0
aiocache==0.1.13
jinja2==2.8
wtforms==2.1

# blog
micawber==0.3.3
markdown==2.6.7
python-slugify==1.2.1
beautifulsoup4==4.5.1

pytest
